Split standard and ZK EVM UInt256 implementations
