+++ src/Directory.Build.targets
@@ -1,0 +1,13 @@
+<Project>
+
+  <ItemGroup Condition="'$(EnableZkEvm)' == 'true'">
+    <Compile Remove="**/*.std.cs" />
+    <None Include="**/*.std.cs" />
+  </ItemGroup>
+
+  <ItemGroup Condition="'$(EnableZkEvm)' != 'true'">
+    <Compile Remove="**/*.zkevm.cs" />
+    <None Include="**/*.zkevm.cs" />
+  </ItemGroup>
+
+</Project>
