+++ src/Nethermind.Int256.Tests/UInt256Tests.cs
@@ -916,13 +916,4 @@ public class UInt256Tests : UInt256TestsTemplate<UInt256>
         }
     }
 
-    [NonParallelizable]
-    [TestCase(false)]
-    [TestCase(true)]
-    public void Shoud_respect_appcontext_switch(bool useHashCodeRandomizer)
-    {
-        AppContext.SetSwitch("Nethermind.Int256.UseHashCodeRandomizer", useHashCodeRandomizer);
-
-        Assert.That(UInt256.UseHashCodeRandomizer, Is.EqualTo(useHashCodeRandomizer));
-    }
 }
